requirements.txt
certifi==2019.6.16
chardet==3.0.4
cloudinary==1.10.0
Django==2.1.4
django-ckeditor==5.6.1
django-fcm==0.1.1
django-js-asset==1.2.2
django-richtextfield==1.3.0
djangorestframework==3.10.2
gunicorn==19.9.0
idna==2.8
mock==3.0.5
paho-mqtt==1.5.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyfcm==1.4.7
python-dateutil==2.7.3
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
